Fix parameter naming in ASR model loading for consistency
- Updated `_load_transformers_pipeline` function in `model_loader.py` to use `torch_dtype` instead of `dtype` for clarity and consistency with the parameter naming convention.

+++ app/asr/model_loader.py
@@ -163,7 +163,7 @@ def _load_transformers_pipeline(settings: Settings) -> Any:
             tokenizer=processor.tokenizer,
             feature_extractor=processor.feature_extractor,
             device=device,
-            dtype=torch_dtype,
+            torch_dtype=torch_dtype,
             chunk_length_s=30,
             return_timestamps=False,
         )
@@ -172,7 +172,7 @@ def _load_transformers_pipeline(settings: Settings) -> Any:
             task="automatic-speech-recognition",
             model=settings.model_id,
             device=device,
-            dtype=torch_dtype,
+            torch_dtype=torch_dtype,
             chunk_length_s=30,
             return_timestamps=False,
         )
